% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_table.R
\name{cont_table}
\alias{cont_table}
\title{Title}
\usage{
cont_table(varlist, data)
}
\arguments{
\item{varlist}{list of variables for table}

\item{data}{data.frame for dataset}
}
\value{
a gtsummary table of class "tbl_summary"
}
\description{
Title
}
\examples{
cont_table(varlist = c("mpg", "wt"),
           data = mtcars)
}
