% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_table.R
\name{cont_table}
\alias{cont_table}
\title{Table of Descriptive Statistics for Continuous Variables}
\usage{
cont_table(varlist, data)
}
\arguments{
\item{varlist}{list of variables for table}

\item{data}{data.frame for dataset}
}
\value{
a gtsummary table of class "tbl_summary"
}
\description{
The \code{cont_table()} function is built upon the
\code{\link[gtsummary:tbl_summary]{gtsummary::tbl_summary()}} function from the \href{https://www.danieldsjoberg.com/gtsummary/index.html}{\code{gtsummary}} package. Try internal package link \link[gtsummary:gtsummary-package]{gtsummary::gtsummary}.
}
\examples{
cont_table(varlist = c("mpg", "wt"), data = mtcars)
}
